% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Graphs accident locations on US state maps}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A valid state number between 1 and 51 as number or integer}

\item{year}{Numerical or integer value that corresponds to a valid year in the data set}
}
\value{
A graphical representation of a select US state with accident locations represented
        as dots on the map
}
\description{
This function plots locations of accidents based on longitude and latitude on select
US state maps for a particular year. The input for this function is a valid state number
\code{state.num} based on the available FARS data, and the \code{year} of interest. Both
values have to be valid numbers/integers for which FARS data is available, otherwise the
function results in an error. This function implements components of the maps package.
}
\details{
This function implements the \code{make_filename()} and \code{fars_read()} function
         of this package.
}
\examples{
\dontrun{fars_map_state(23, 2013)}
\dontrun{fars_map_state(1, 2014)}
\dontrun{fars_map_state(51, 2015)}

#Errors will result in following cases:
\dontrun{fars_map_state(23, 2018)}
\dontrun{fars_map_state(52, 2013)}

}
