% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads accident file and returns tibble with month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A single year, or a vector/list of multiple \code{years} in numerical/integer class}
}
\value{
A tibble of each valid year that was chosen as input, diplaying "Month" and "year" of the
        accident file.
}
\description{
This function takes one year, a vector of \code{years}, or a list of \code{years} as input for
which accident files are available. When accident files are not available the function will result
in a warning message indicating that the year is not valid. For valid input the function will
return a tibble that displays "Month" and "year" of the accident file.
}
\details{
This function implements the \code{make_filename()} and \code{fars_read()} function of this
         package
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years(c(2013, 2014, 2015))}

}
