% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Returns the number of accidents per month of select years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A single year, or a vector/list of multiple \code{years} in numerical/integer class}
}
\value{
A summary in tibble format for each valid year, indicating the number of total accidents
        in a month for each month for the whole year.
}
\description{
This function takes one year, a vector of \code{years}, or a list of \code{years} as input for
which accident files are available. Years for which no accident file is available will result in
an error and a worning message. For years that accident files are available the function summarizes
the accident data for each year based on "Month" and the corresponding total number of accidents for
that month. This function utilizes dplyr and tidyr, and returns the results as tibble.
}
\details{
This function implements the \code{fars_read_years()}, \code{make_filename()} and
         \code{fars_read()} function of this package.
}
\examples{
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(c(2013, 2014, 2015))}

}
