% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename according to the input year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numerical or integer value that corresponds to a year in the data set}
}
\value{
A character string with the pattern "accident_%d.csv.bz2", where "%d" is replaced with
a number/integer that corresponds to a year in the data set (i.e "2013", "2014", or "2015").
This output character string can be used as \code{filename} input for the fars_read function.
}
\description{
This function creates a filename with the pattern "accident_%d.csv.bz2", where the "%d" is
replaced by the numerical or intiger function input \code{year}. Input that is not numerical
or an integer will lead to an error.
}
\examples{
\dontrun{make_filename(2013)}
\dontrun{fars_read(make_filename(2013))}

}
